infra/auto-stop-vm/requirements.txt
functions-framework==3.*
google-cloud-compute==1.*
